Στάδιο ολοκλήρωσης αναμνηστικού

+++ app/src/main/java/com/kasal/podoapp/data/VisitDao.kt
@@ -26,7 +26,6 @@ interface VisitDao {
     """)
     fun forPatient(patientId: Int): Flow<List<Visit>>
 
-    // Για προβολή ημερολογίου επισκέψεων: εύρος ημέρας [start, end] (UTC millis)
     @Query("""
         SELECT * FROM visits
         WHERE dateTime BETWEEN :start AND :end
@@ -34,7 +33,6 @@ interface VisitDao {
     """)
     suspend fun getVisitsForDate(start: Long, end: Long): List<Visit>
 
-    // (Προαιρετικό) Αν χρειαστεί να ανοίγεις VisitDetail με id
     @Query("SELECT * FROM visits WHERE id = :id LIMIT 1")
-    suspend fun getById(id: Long): Visit?
+    suspend fun getById(id: Int): Visit?
 }
